services/ai-orchestrator/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
openai>=1.0.0
python-dotenv>=1.0.0
redis[hiredis]>=5.0.0